Merge to frontend

+++ backend/src/main/java/com/smarthire/resume/service/CandidatoService.java
@@ -15,10 +15,13 @@ import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @Service
@@ -64,17 +67,36 @@ public class CandidatoService {
         );
     }
 
-    public Candidato buscarCandidatoPorNome(String nomeCandidato) {
-        return candidatoRepository.findByNome(nomeCandidato)
-                .orElseThrow(() -> new BusinessRuleException("Candidato não encontrado."));
+    public List<CandidatoDto> listarTodos() {
+        List<Candidato> candidatos = candidatoRepository.findAll();
+        return candidatos.stream()
+                .map(this::listarCandidatos)
+                .collect(Collectors.toList());
     }
 
+    public List<CandidatoDto> buscarCandidatoPorNome(String nomeCandidato) {
+        List<Candidato> candidatos = candidatoRepository.findByNomeContainingIgnoreCase(nomeCandidato);
+        List<CandidatoDto> candidatosDto = candidatos.stream()
+                .map(this::listarCandidatos)
+                .collect(Collectors.toList());
+        if (candidatosDto.isEmpty()) {
+            throw new BusinessRuleException("Candidato não encontrado.");
+        }
+        return candidatosDto;
+    }
+
+    @Transactional
     public void deletarCandidatoPorId(UUID id){
+        if (!candidatoRepository.existsById(id)) {
+            throw new BusinessRuleException("Candidato não encontrado.");
+        }
         Candidato candidato = candidatoRepository.findById(id)
                 .orElseThrow(() -> new BusinessRuleException("Candidato não encontrado."));
         candidatoRepository.delete(candidato);
     }
 
+
+    @Transactional
     public void criarComCurriculo(Curriculo curriculo) {
         Candidato candidato = new Candidato();
         candidato.setCurriculo(curriculo);
@@ -85,8 +107,10 @@ public class CandidatoService {
 
         salvar(candidato);
     }
-
     public void adicionarCandidatoAVaga(UUID idCandidato, UUID idVaga) {
+        if (!candidatoRepository.existsById(idCandidato) || !vagaRepository.existsById(idVaga)) {
+            throw new BusinessRuleException("Candidato ou vaga não encontrado.");
+        }
         Candidato candidato = candidatoRepository.findById(idCandidato)
                 .orElseThrow(() -> new BusinessRuleException("Candidato não encontrado."));
         Vaga vaga = vagaRepository.findById(idVaga)
